playwright_py/requirements.txt
pytest>=8.0
pytest-playwright>=0.5.0,<1.0
playwright>=1.48,<2.0
pytest-html>=4.1.1
python-dotenv>=1.0.1
